// apps/desktop/packages/shared/generated/ArtifactSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ArtifactId } from "./ArtifactId.js";
import type { ArtifactKind } from "./ArtifactKind.js";
import type { RunId } from "./RunId.js";

export type ArtifactSummary = { id: ArtifactId, runId: RunId, kind: ArtifactKind, storagePath: string, };